fix: 🐛 NPE when finishing Snyk Wizard in empty workspace
SnykWizard was triggering authentication and trust on finishing without checking
the workspace for projects. This leads to a NPE, as our language server client
is still uninitialized. LSP4e only initializes the language client and starts the
language server, when a project / file is associated with the corresponding
mime/media type.

This fix wraps authentication and trust into a condition, checking, if the workspace
contains projects.